feat(globing): added globing module
at the moment the implementation of this module come too late ; the substituted token is not interpreted as it should be. The substitution has to be done upstream ; between the tokenizer and the AST.
the right implementation will be fixed soon thanks to a teamwork between bprunevi and baavril.

+++ src/expansions/maths_exp/parentheses.c
@@ -64,5 +64,5 @@ int			ft_parentheses_nbr(char *expr)
 			brk_check--;
 		i++;
 	}
-	return ((par_check) || (brk_check) ? -1 : par_nbr);
+	return (((par_check) || (brk_check)) ? -1 : (int)par_nbr);
 }
